# optimize ossl_sm4_set_key speed