this optimization comes from libgcrypt, increses about 48% speed.

Benchmark on my AMD Ryzen Threadripper 3990X.

before:
Did 5752000 SM4 setup operations in 1000151us (5751131.6 ops/sec) 
after:
Did 8506000 SM4 setup operations in 1000023us (8505804.4 ops/sec)

<!--
Thank you for your pull request. Please review these requirements:

Contributors guide: https://github.com/openssl/openssl/blob/master/CONTRIBUTING.md

Other than that, provide a description above this comment if there isn't one already

If this fixes a GitHub issue, make sure to have a line saying 'Fixes #XXXX' (without quotes) in the commit message.
-->

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->
- [ ] documentation is added or updated
- [ ] tests are added or updated